WIP refactor Editor in DO.U
